refactor(mobile): rewrite intra-feature imports under Signer/
